Version 0.0.4

+++ TASKER/db/user_db.py
@@ -2,12 +2,15 @@ from sqlalchemy.future import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy import and_, or_
+from datetime import datetime
 from fastapi.responses import JSONResponse
-from fastapi import HTTPException, status
+from fastapi import status
 from TASKER.core.security import hash_password, generate_token
 from TASKER.api.schemas.users import Login, StatusFriend
 from TASKER.api.schemas.utils import SearchRequest
-from TASKER.db.models import UserDB, FriendshipDB, FriendRequestDB
+from TASKER.core.utils import get_list_user
+from TASKER.db.models import UserDB, FriendshipDB, FriendRequestDB, NotificationDB
+from typing import Dict
 import logging
 
 
@@ -16,8 +19,8 @@ async def user_registration(data: Login, db: AsyncSession):
     user = await db.execute(statement)
     user = user.scalar_one_or_none()
     if user:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail='Нікнейм зайнятий, оберіть інший')
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST, content='Нікнейм зайнятий, оберіть інший')
 
     data.password = hash_password(data.password)
     new_user = UserDB(username=data.username, password=data.password)
@@ -29,8 +32,8 @@ async def user_registration(data: Login, db: AsyncSession):
     except IntegrityError:
         logging.error('user_registration-Integrity', exc_info=True)
         await db.rollback()
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
-                            detail='Помилка при реєстрації')
+        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                            content='Помилка при реєстрації')
 
     token = generate_token(new_user)
     response = JSONResponse(status_code=status.HTTP_201_CREATED, content='')
@@ -43,11 +46,11 @@ async def user_login(data: Login, db: AsyncSession):
     user = await db.execute(statement)
     user = user.scalar_one_or_none()
     if not user:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail='Невірний пароль або логін')
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST, content='Невірний пароль або логін')
     if user.password != hash_password(data.password):
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail='Невірний пароль або логін')
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST, content='Невірний пароль або логін')
 
     token = generate_token(user)
     response = JSONResponse(status_code=status.HTTP_200_OK, content='')
@@ -57,8 +60,8 @@ async def user_login(data: Login, db: AsyncSession):
 
 async def user_send_request_friend(username: str, friend_name: str,  db: AsyncSession):
     if username == friend_name:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail='Не можна відправити собі запит')
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST, content='Не можна відправити собі запит')
 
     statement_check_friendship = select(FriendRequestDB).filter(and_(
         FriendRequestDB.sender_name == username,
@@ -68,8 +71,8 @@ async def user_send_request_friend(username: str, friend_name: str,  db: AsyncSe
     check_friendship = check_friendship.scalar_one_or_none()
     if check_friendship:
         if check_friendship.status in (StatusFriend.pending.value, StatusFriend.accepted.value):
-            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=(
-                                'Запит вже надіслано', 'Вже у друзях')[check_friendship.status == StatusFriend.accepted.value])
+            return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST, content=(
+                'Запит вже надіслано', 'Вже у друзях')[check_friendship.status == StatusFriend.accepted.value])
 
     statement = select(UserDB).filter(
         and_(UserDB.username.in_([username, friend_name])))
@@ -77,28 +80,36 @@ async def user_send_request_friend(username: str, friend_name: str,  db: AsyncSe
     users = users.scalars().all()
 
     if len(users) != 2:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail='Профіль не знайдено')
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST, content='Профіль не знайдено')
     user, friend = users
 
     if user.username == friend_name:
         friend, user = user, friend
     try:
+        # Створюємо запит
         request = FriendRequestDB(
             sender_name=user.username, receiver_name=friend.username)
         db.add(request)
         await db.commit()
+        await db.refresh(request)
+
+        # Створюємо повідомлення
+        notification = NotificationDB(
+            username=friend_name, friend_request_id=request.id)
+        db.add(notification)
+        await db.commit()
         return JSONResponse(status_code=status.HTTP_200_OK, content='Надіслано')
     except IntegrityError:
         await db.rollback()
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
-                            detail='Помилка додавання запиту на дружбу до бази даних')
+        return JSONResponse(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                            content='Помилка додавання запиту на дружбу')
 
 
 async def accept_friend_request(username: str, friend_name: str, db: AsyncSession):
     if username == friend_name:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail='Не можна прийняти свій запит')
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST, content='Не можна прийняти свій запит')
 
     statement = select(FriendRequestDB).filter(
         and_(FriendRequestDB.sender_name == friend_name,
@@ -110,12 +121,12 @@ async def accept_friend_request(username: str, friend_name: str, db: AsyncSessio
         friend_request = friend_request.scalar_one_or_none()
     except IntegrityError:
         await db.rollback()
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail='Помилка сервера')
+        return JSONResponse(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, content='Помилка сервера')
 
     if not friend_request:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail='Запит не знайдено')
+        return JSONResponse(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, content='Запит не знайдено')
 
     friend_request.status = 'accepted'
     friendship = FriendshipDB(
@@ -123,18 +134,32 @@ async def accept_friend_request(username: str, friend_name: str, db: AsyncSessio
 
     db.add(friendship)
     try:
+        # Створюємо "Дружбу"
         await db.commit()
+        await db.refresh(friend_request)
+        # Помічаемо повідомлення як прочитане
+        statement = select(NotificationDB).where(and_(
+            NotificationDB.username == username,
+            NotificationDB.friend_request_id == friend_request.id))
+        notification = await db.execute(statement)
+        notification = notification.scalar_one_or_none()
+        if notification:
+            notification.is_read = True
+
+        db.add(notification)
+        await db.commit()
+
         return JSONResponse(status_code=status.HTTP_201_CREATED, content='Додано до друзів')
     except IntegrityError:
         await db.rollback()
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail='Помилка сервера')
+        return JSONResponse(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, content='Помилка сервера')
 
 
 async def declain_friend_request(username: str, friend_name: str, db: AsyncSession):
     if username == friend_name:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail='Ви не надсилали собі запит :)')
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST, content='Ви не надсилали собі запит :)')
 
     statement = select(FriendRequestDB).where(or_(and_(
         FriendRequestDB.sender_name == username,
@@ -147,18 +172,32 @@ async def declain_friend_request(username: str, friend_name: str, db: AsyncSessi
     try:
         friend_request = await db.execute(statement)
         friend_request = friend_request.scalar_one_or_none()
+
         if not friend_request:
-            raise HTTPException(status.HTTP_404_NOT_FOUND,
-                                detail='Запит не занйдено')
+            return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST, content='Запит не знайдено')
+
+        # Помічаемо повідомлення як прочитане
+        statement = select(NotificationDB).where(and_(
+            NotificationDB.username == username,
+            NotificationDB.friend_request_id == friend_request.id))
+        notification = await db.execute(statement)
+        notification = notification.scalar_one_or_none()
+        if not notification:
+            return JSONResponse(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, content='Помилка сервера')
+        notification.is_read = True
+
+        db.add(notification)
         await db.delete(friend_request)
         await db.commit()
+
         return JSONResponse(status_code=status.HTTP_200_OK, content='Запит відхилено')
 
     except:
         await db.rollback()
         logging.error(msg='declain_friend_request', exc_info=True)
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail='Помилка сервера')
+        return JSONResponse(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, content='Помилка сервера')
 
 
 async def delete_friendship(username: str, friend_name: str, db: AsyncSession):
@@ -173,8 +212,8 @@ async def delete_friendship(username: str, friend_name: str, db: AsyncSession):
     friendship = await db.execute(statement)
     friendship = friendship.scalar_one_or_none()
     if not friendship:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Зв'язок не знайдено")
+        return JSONResponse(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, content="Зв'язок не знайдено")
 
     statement2 = select(FriendRequestDB).filter(or_(and_(
         FriendRequestDB.sender_name == username,
@@ -193,11 +232,11 @@ async def delete_friendship(username: str, friend_name: str, db: AsyncSession):
         await db.commit()
         return JSONResponse(status_code=status.HTTP_200_OK, content='Видалено з друзів')
     except IntegrityError:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail='Помилка сервера')
+        return JSONResponse(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, content='Помилка сервера')
 
 
-async def get_list_friends(token: str, db: AsyncSession):
+async def get_list_friends(token: Dict, db: AsyncSession):
     statement = select(FriendshipDB).filter(or_(and_(
         FriendshipDB.user_name == token['username']),
         and_(
@@ -214,24 +253,36 @@ async def get_search(data: SearchRequest, token: dict, db: AsyncSession):
         UserDB.username.ilike(f'%{data.request}%'),
         UserDB.username.ilike(f'{data.request}%'),
         UserDB.username.ilike(f'%{data.request}')),
-        and_(UserDB.id != token['id'])).limit(15)
+        and_(UserDB.id != token['id'])).order_by('username').limit(15)
 
     result_user = await db.execute(statement_user)
     result_user = result_user.scalars().all()
-    # Поля які потрібно повернути
-    fields = ['id', 'username', 'online', 'last_seen']
 
     if result_user:
-        list_user = []
-        # list_friends = await get_list_friends(token=token, db=db)
-        for user in result_user:
-            user_dict = user.as_dict(fields=fields)
-            # Отримуємо список ідентифікаторів друзів користувача
-            friend_ids = [friend.id for friend in user.friends]
-            # Перевіряємо, чи є користувач у друзях
-            user_dict['is_friend'] = token['id'] in friend_ids
-            list_user.append(user_dict)
+        list_user = get_list_user(result_user, token)
 
-        return list_user
-    response = JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content='')
+        response = JSONResponse(
+            status_code=status.HTTP_200_OK, content=list_user)
+        return response
+    return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content='')
+
+
+async def logout(request):
+    response = JSONResponse(status_code=status.HTTP_200_OK, content='')
+    response.delete_cookie(key='TOPUS')
     return response
+
+
+async def set_user_status(user_id: int, online: bool, db: AsyncSession):
+    try:
+        user = await db.get(UserDB, user_id)
+        if user:
+            if not online:
+                user.last_seen = datetime.now()
+            user.online = online
+            await db.commit()
+        else:
+            return JSONResponse(
+                status_code=status.HTTP_404_NOT_FOUND, content='Юзера не знайдено')
+    except:
+        logging.error(msg='set_user_status', exc_info=True)
